refactor: merge setup/docker/explicit/scripts/lib into core/lib
