+++ core/lib/log/buildkitd-log-parser.test.ts
@@ -1,5 +1,5 @@
-import { describe, it, assert, reportResults } from "../../../../../core/shared/test/test-shim.js";
-import { parseEntries, parseDenialTimeline } from "./buildkitd-log-parser.js";
+import { describe, it, assert, reportResults } from "../test/test-shim.ts";
+import { parseEntries, parseDenialTimeline } from "./buildkitd-log-parser.ts";
 
 // Real line captured from a live moby/buildkit v0.31.1 explicit-mode container.
 const REAL_DENY_LINE =
@@ -74,7 +74,7 @@ describe("parseDenialTimeline", () => {
   });
 });
 
-// aggregate() itself is tested in core/shared/lib/aggregate.test.ts.
-// parseIdentifier() itself is tested in core/shared/lib/parse-identifier.test.ts.
+// aggregate() itself is tested in core/lib/log/aggregate.test.ts.
+// parseIdentifier() itself is tested in core/lib/log/parse-identifier.test.ts.
 
 reportResults();
